// apps/frontend/src/types/theme.types.ts
export interface ICorePalette {
  // Special
  blank: string;

  // Neutrals
  offBlack200: string;
  offBlack100: string;
  grey200: string;
  grey100: string;
  grey50: string;
  grey30: string;
  grey10: string;
  white: string;

  // Blue
  blue300: string;
  blue200: string;
  blue150: string;
  blue100: string;
  blue50: string;
  blue30: string;
  blue10: string;

  // Turquoise
  turquoise300: string;
  turquoise200: string;
  turquoise150: string;
  turquoise100: string;
  turquoise50: string;
  turquoise30: string;
  turquoise10: string;

  // Green
  green300: string;
  green200: string;
  green150: string;
  green100: string;
  green50: string;
  green30: string;
  green10: string;

  // Red
  red300: string;
  red200: string;
  red150: string;
  red100: string;
  red50: string;
  red30: string;
  red10: string;

  // Magenta
  magenta300: string;
  magenta200: string;
  magenta150: string;
  magenta100: string;
  magenta50: string;
  magenta30: string;
  magenta10: string;

  // Violet
  violet300: string;
  violet200: string;
  violet150: string;
  violet100: string;
  violet50: string;
  violet30: string;
  violet10: string;

  // Purple
  purple300: string;
  purple200: string;
  purple150: string;
  purple100: string;
  purple50: string;
  purple30: string;
  purple10: string;

  // Orange
  orange300: string;
  orange200: string;
  orange150: string;
  orange100: string;
  orange50: string;
  orange30: string;
  orange10: string;

  // Yellow
  yellow300: string;
  yellow200: string;
  yellow150: string;
  yellow100: string;
  yellow50: string;
  yellow30: string;
  yellow10: string;

  // Warm Red
  warmRed300: string;
  warmRed200: string;
  warmRed150: string;
  warmRed100: string;
  warmRed50: string;
  warmRed30: string;
  warmRed10: string;
}

export interface IChipColours {
  // Blue
  chipBlueBg: string;
  chipBlueText: string;

  // Green
  chipGreenBg: string;
  chipGreenText: string;

  // Grey
  chipGreyBg: string;
  chipGreyText: string;

  // Disabled
  chipDisabledBg: string;
  chipDisabledText: string;

  // Violet
  chipVioletBg: string;
  chipVioletText: string;

  // Purple
  chipPurpleBg: string;
  chipPurpleText: string;

  // Magenta
  chipMagentaBg: string;
  chipMagentaText: string;

  // Red
  chipRedBg: string;
  chipRedText: string;

  // Orange
  chipOrangeBg: string;
  chipOrangeText: string;
}

export interface ISignificanceColours {
  // Red
  red100: string;
  red70: string;
  red50: string;
  red40: string;
  red30: string;
  red20: string;
  red10: string;
  red0: string;

  // Neutral
  neutral: string;

  // Green
  green0: string;
  green10: string;
  green20: string;
  green30: string;
  green40: string;
  green50: string;
  green70: string;
  green100: string;

  // Violet
  violet50: string;
}